Refactor plugins and modules
