+++ src/pages/community/index.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import { CEM } from '~/plugins/custom-event-manager'
+import { CEM } from '~/modules/custom-event-manager'
 import pageBase from '~/mixins/page-base'
 import PostItem from '~/components/community/PostItem'
