+++ src/pages/vacant/_buildingId/_floorId/index.vue
@@ -151,7 +151,7 @@ import axios from 'axios'
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock'
 import pageBase from '~/mixins/page-base'
 import modalScroll from '~/mixins/modal-scroll'
-import ExpireCounter from '~/plugins/ExpireCounter'
+import ExpireCounter from '~/modules/expire-counter'
 import { Grid, ArrowBlock } from '~/components/ui'
 
 export default {
